FIRST VERSION WORKING !

+++ include/codage_rle.h
@@ -18,6 +18,6 @@ typedef struct {
     int profondeur;
 } Chemin_Huff;
 
-void chaine_Huff_vect(FILE *f, int *coeffs, bool is_Y, bool is_Cb, int predicateur);
+void chaine_Huff_vect(FILE *f, int16_t *coeffs, bool is_Y, bool is_Cb, int predicateur);
 
 // #endif
